Rename switchlink_link.h to switchlink_link_types.h
